# backend-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "STAY STRONG API",
        "description": "API para gestión de programas de entrenamiento",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/training-sessions/{id}": {
            "get": {
                "tags": [
                    "TrainingSessions"
                ],
                "summary": "Obtener una sesión de entrenamiento específica",
                "operationId": "cf90a0aef342a8ec382d7c47a0b6309a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión de entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesión encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrainingSession"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "TrainingSessions"
                ],
                "summary": "Actualizar una sesión de entrenamiento",
                "operationId": "256b8ee0ff199d6e6364973e6afa5623",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión de entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TrainingSessionUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión actualizada correctamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TrainingSession"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/training-sessions/{session_id}/exercises/{exercise_id}": {
            "patch": {
                "tags": [
                    "TrainingSessions"
                ],
                "summary": "Actualizar un ejercicio de sesión",
                "operationId": "4c69ed47fa427e24a838c2bff2923e23",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "description": "ID de la sesión de entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "exercise_id",
                        "in": "path",
                        "description": "ID del ejercicio de la sesión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SessionExerciseUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ejercicio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ejercicio actualizado correctamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/SessionExercise"
                                        },
                                        "session_status": {
                                            "type": "string",
                                            "example": "completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El ejercicio no pertenece a esta sesión"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sesión o ejercicio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/training-sessions/{id}/exercises": {
            "get": {
                "tags": [
                    "TrainingSessions"
                ],
                "summary": "Listar ejercicios de una sesión",
                "operationId": "f019326602fe178deb366d83ce723de6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sesión de entrenamiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ejercicios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "session_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "session_status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "exercises": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SessionExercise"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sesión no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TrainingSession": {
                "title": "Training Session",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "program_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ],
                        "example": "pending"
                    },
                    "comments": {
                        "type": "string",
                        "example": "Buena sesión",
                        "nullable": true
                    },
                    "estimated_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "program": {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "session_exercises": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SessionExercise"
                        }
                    }
                },
                "type": "object"
            },
            "SessionExercise": {
                "title": "Session Exercise",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "training_session_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "exercise_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lifted_weight": {
                        "type": "integer",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ],
                        "example": "completed"
                    },
                    "exercise": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Press de banca"
                            },
                            "muscle_group": {
                                "type": "string",
                                "example": "Pecho"
                            },
                            "series": {
                                "type": "integer",
                                "example": 4
                            },
                            "repetitions": {
                                "type": "integer",
                                "example": 12
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TrainingSessionUpdate": {
                "title": "Training Session Update",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ],
                        "example": "completed"
                    },
                    "comments": {
                        "type": "string",
                        "example": "Buena sesión",
                        "nullable": true
                    },
                    "estimated_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    }
                },
                "type": "object"
            },
            "SessionExerciseUpdate": {
                "title": "Session Exercise Update",
                "properties": {
                    "lifted_weight": {
                        "type": "integer",
                        "example": 55
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ],
                        "example": "completed"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Usar el token JWT obtenido al autenticarse"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}